Further fixes

+++ users-delete.php
@@ -1,16 +1,19 @@
 <?php
 	include('rights.php');
 	
+	if($_GET['id'] == "") {
+		$_SESSION['error'] = 'No ID given.';
+		exit(header('Location: users-manage.php'));
+	}
 	$id = $_GET['id'];
 	
 	include 'MySQLCredentials.php';
 	
 	if($connection->query("DELETE from $users WHERE user_id='$id'")) {
-		echo "Record deleted.";
+		$_SESSION['success'] = "Record deleted.";
 	} else {
-		echo $connection->errorno . ": " . $connection->error . "\n";
-		$connection->close();
-		exit;
+		$_SESSION['error'] = $connection->errno . ": " . $connection->error . "\n";
 	}
 	$connection->close(); // Closing Connection
+	exit(header('Location: users-manage.php'));
 ?>
